+++ scripts/deploy-github-release.ps1
@@ -1,5 +1,5 @@
 ﻿param(
-    [string]$Version = "1.2.1",
+    [string]$Version = "1.2.2",
     [string]$Repository = "Xian-JL/dsh-Kinich-theme"
 )